requirements.txt
discord.py==1.7.3
Pillow==9.0.1
ttf-opensans==2020.10.30